feature Player Edition

+++ app/src/main/java/com/decathlon/canaveral/common/model/BaseItem.kt
@@ -3,6 +3,6 @@ package com.decathlon.canaveral.common.model
 import android.os.Parcelable
 
 abstract class BaseItem : Parcelable {
-    abstract val id: Int
+    abstract val id: Long
 
 }
